CRM457-1328: Submission notification emails are retryable (#728)
## Description of change
- The notification email is handled as its own separate job so that if
it fails it will be retried without retriggering app store submission
- Clear out the old SNS stuff and a redundant state update in
SubmitToAppStore
- Make SubmitToAppStore a standard job so that the now-vs-later logic is
handled by our standard `RUN_SIDEKIQ_IN_TEST_MODE` logic without an
extra layer of complexity.

[Link to relevant
ticket](https://dsdmoj.atlassian.net/browse/CRM457-1328)

---------
