+++ spec/system/nsm/solicitor_declaration_spec.rb
@@ -2,10 +2,9 @@ require 'rails_helper'
 
 RSpec.describe 'User can fill in solicitor declaration', type: :system do
   let(:claim) { create(:claim) }
-  let(:app_store_notifier) { instance_double(SubmitToAppStore, process: true) }
 
   before do
-    allow(SubmitToAppStore).to receive(:new).and_return(app_store_notifier)
+    allow(SubmitToAppStore).to receive(:perform_later)
     visit provider_saml_omniauth_callback_path
   end
 
@@ -20,7 +19,6 @@ RSpec.describe 'User can fill in solicitor declaration', type: :system do
       signatory_name: 'John Doe',
       status: 'submitted'
     )
-    expect(SubmitToAppStore).to have_received(:new)
-    expect(app_store_notifier).to have_received(:process)
+    expect(SubmitToAppStore).to have_received(:perform_later)
   end
 end
