+++ spec/system/prior_authority/check_your_answers/sent_back_for_correct_info_spec.rb
@@ -4,6 +4,7 @@ RSpec.describe 'Prior authority applications, sent back for info correction - ch
   before do
     visit provider_saml_omniauth_callback_path
     visit prior_authority_steps_check_answers_path(application)
+    allow(SubmitToAppStore).to receive(:perform_later)
   end
 
   let(:application) do
